+++ internal/ensresolver/repo.go
@@ -1,0 +1,33 @@
+package ensresolver
+
+import (
+	"time"
+
+	"gorm.io/gorm"
+	"gorm.io/gorm/clause"
+)
+
+const (
+	defaultBatchSize = 100
+)
+
+type Repo struct {
+	db *gorm.DB
+}
+
+func NewRepo(db *gorm.DB) *Repo {
+	return &Repo{db: db}
+}
+
+type EnsName struct {
+	CreatedAt time.Time
+	UpdatedAt time.Time
+	Address   string
+	Name      string
+}
+
+func (r *Repo) BatchCreate(data []EnsName) error {
+	return r.db.Model(&EnsName{}).Clauses(clause.OnConflict{
+		DoNothing: true,
+	}).CreateInBatches(data, defaultBatchSize).Error
+}